Handle models with fewer that required waveforms and waveforms that stop early
For accuracy, we need two waveforms for a model with a pullup and pulldown,
however, some IBIS files don't have that. Produce a kpu/kpd pair with only
one set of waveforms, it won't be very good, but it won't raise an exception
and prevent the whole file from being processed

Some fixtures have waveforms that stop in time before other waveforms in the model.
Just extend out the ones that stop early so they cover the same time period.